requirements.txt
discord.py[voice]==1.7.2
wavelink==0.9.9
pip==21.1.1
setuptools==56.1.0
